= http: add test for cookie with commas in value
Due to #869, a comma is a valid separator, and so everything after the
first comma in a cookie's value (there shouldn't be commas in the cookie
value, but a user agent is required to be tolerant of them according to
RFC6265) is silently discarded.

This isn't a breaking change because without this fix, any cookie
containing a comma in the value would result in *all* the cookies being
discarded (albeit with a log message, which now will not appear).

Correct behaviour would require not allowing commas as a separator.